Move fetch functions into api and http modules
