+++ __tests__/Auth0Client/getIdTokenClaims.test.ts
@@ -14,7 +14,7 @@ import { TEST_CODE_CHALLENGE } from '../constants';
 jest.mock('unfetch');
 jest.mock('es-cookie');
 jest.mock('../../src/jwt');
-jest.mock('../../src/token.worker');
+jest.mock('../../src/worker/token.worker');
 
 const mockWindow = <any>global;
 const mockFetch = (mockWindow.fetch = <jest.Mock>unfetch);
