+++ __tests__/Auth0Client/getTokenWithPopup.test.ts
@@ -22,7 +22,7 @@ import { Auth0ClientOptions } from '../../src';
 jest.mock('unfetch');
 jest.mock('es-cookie');
 jest.mock('../../src/jwt');
-jest.mock('../../src/token.worker');
+jest.mock('../../src/worker/token.worker');
 
 const mockWindow = <any>global;
 const mockFetch = (mockWindow.fetch = <jest.Mock>unfetch);
